+++ Graphics/Render3D.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Diagnostics;
-using DWORD = System.Int16;
+using DWORD = System.UInt32;
 
 namespace FalconNet.Graphics
 {
